app-backend/pyproject.toml
[tool.poetry]
name = "app-backend"
version = "0.1.0"
description = "Backend for football analytics app"
authors = ["chonalchendo <110059232+chonalchendo@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
pandas = "^2.2.2"
joblib = "^1.4.2"
google-cloud-storage = "^2.16.0"
rich = "^13.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
